lab4/requirements.txt
numpy
scipy
pyqtgraph
PySide6
pyopengl
matplotlib